e2e test cases refactor
