+++ test/e2e/channel_single_event_test.go
@@ -19,21 +19,11 @@ limitations under the License.
 package e2e
 
 import (
-	"fmt"
 	"testing"
 
-	"k8s.io/apimachinery/pkg/util/uuid"
 	"knative.dev/eventing/test/base/resources"
-	"knative.dev/eventing/test/common"
+	"knative.dev/eventing/test/e2e/helpers"
 )
-
-func TestSingleBinaryEventForChannel(t *testing.T) {
-	SingleEventForChannelTestHelper(t, resources.CloudEventEncodingBinary, channels)
-}
-
-func TestSingleStructuredEventForChannel(t *testing.T) {
-	SingleEventForChannelTestHelper(t, resources.CloudEventEncodingStructured, channels)
-}
 
 /*
 SingleEventForChannelTestHelper tests the following scenario:
@@ -41,54 +31,19 @@ SingleEventForChannelTestHelper tests the following scenario:
 EventSource ---> Channel ---> Subscription ---> Service(Logger)
 
 */
-func SingleEventForChannelTestHelper(t *testing.T, encoding string, channels []string) {
-	runTests(t, channels, common.FeatureBasic, func(st *testing.T, channel string) {
-		channelName := "e2e-singleevent-channel-" + encoding
-		senderName := "e2e-singleevent-sender-" + encoding
-		subscriptionName := "e2e-singleevent-subscription-" + encoding
-		loggerPodName := "e2e-singleevent-logger-pod-" + encoding
 
-		st.Logf("Run test with channel %q", channel)
-		client := setup(st, true)
-		defer tearDown(client)
+func TestSingleBinaryEventForChannel(t *testing.T) {
+	helpers.SingleEventForChannelTestHelper(
+		t,
+		resources.CloudEventEncodingBinary,
+		channelTestRunner,
+	)
+}
 
-		// create channel
-		channelTypeMeta := getChannelTypeMeta(channel)
-		client.CreateChannelOrFail(channelName, channelTypeMeta)
-
-		// create logger service as the subscriber
-		pod := resources.EventLoggerPod(loggerPodName)
-		client.CreatePodOrFail(pod, common.WithService(loggerPodName))
-
-		// create subscription to subscribe the channel, and forward the received events to the logger service
-		client.CreateSubscriptionOrFail(
-			subscriptionName,
-			channelName,
-			channelTypeMeta,
-			resources.WithSubscriberForSubscription(loggerPodName),
-		)
-
-		// wait for all test resources to be ready, so that we can start sending events
-		if err := client.WaitForAllTestResourcesReady(); err != nil {
-			st.Fatalf("Failed to get all test resources ready: %v", err)
-		}
-
-		// send fake CloudEvent to the channel
-		body := fmt.Sprintf("TestSingleEvent %s", uuid.NewUUID())
-		event := &resources.CloudEvent{
-			Source:   senderName,
-			Type:     resources.CloudEventDefaultType,
-			Data:     fmt.Sprintf(`{"msg":%q}`, body),
-			Encoding: encoding,
-		}
-
-		if err := client.SendFakeEventToAddressable(senderName, channelName, channelTypeMeta, event); err != nil {
-			st.Fatalf("Failed to send fake CloudEvent to the channel %q", channelName)
-		}
-
-		// verify the logger service receives the event
-		if err := client.CheckLog(loggerPodName, common.CheckerContains(body)); err != nil {
-			st.Fatalf("String %q not found in logs of logger pod %q: %v", body, loggerPodName, err)
-		}
-	})
+func TestSingleStructuredEventForChannel(t *testing.T) {
+	helpers.SingleEventForChannelTestHelper(
+		t,
+		resources.CloudEventEncodingStructured,
+		channelTestRunner,
+	)
 }
