+++ test/e2e/helpers/README.md
@@ -1,0 +1,4 @@
+# helpers
+
+This package contains helper functions that are used in the actual tests.
+These functions can be referenced in other repositories (like `eventing-contrib`) for adding similar tests.
